requirements.txt
Flask
python-dotenv
PyMuPDF
requests
google-generativeai
Werkzeug
